Change landing page
Separated the login page and landing page by using @aws-vue/ui-component library which toggles router depending on the state of the user. If user is in logged out state, the div element will display login text and if logged in the div element will display profile text. When user clicks on the router, the profile page will be routed which is locked behind a amplify authenticate tag which will display login screen or profile depending on the state.