update on Authenticate
There was a major work done with authenticator and datastore. When I set up the app there wasn't any ways to store the data to specific user's bucket so I set that up but then I realized authenticator has to be set up on both the app and the profile differently so that the server knows that the current user is requesting their profile. Long story short the now the data is stored to the unique user and it can be accessed by the same user. One glitch I still am working on is that once you are logged in, if you change the url path to a different user name it will show the different user's information which is definitely not ideal for security. I will have to work on that before I can commit this to master branch.